hub/services/hub_status_upgrade_test.go
package services_test

import (
	"errors"
	"io/ioutil"
	"os"
	"path/filepath"

	"github.com/greenplum-db/gp-common-go-libs/testhelper"
	"github.com/greenplum-db/gpupgrade/hub/services"
	"github.com/greenplum-db/gpupgrade/hub/upgradestatus"
	pb "github.com/greenplum-db/gpupgrade/idl"
	"github.com/greenplum-db/gpupgrade/testutils"
	"github.com/greenplum-db/gpupgrade/utils"
	"golang.org/x/net/context"

	"google.golang.org/grpc"

	. "github.com/onsi/ginkgo"
	. "github.com/onsi/gomega"
)

var _ = Describe("status upgrade", func() {
	var (
		hub                      *services.Hub
		fakeStatusUpgradeRequest *pb.StatusUpgradeRequest
		dir                      string
		mockAgent                *testutils.MockAgentServer
		clusterPair              *utils.ClusterPair
		testExecutor             *testhelper.TestExecutor
		cm                       *testutils.MockChecklistManager
	)

	BeforeEach(func() {
		var port int
		mockAgent, port = testutils.NewMockAgentServer()
		mockAgent.StatusConversionResponse = &pb.CheckConversionStatusReply{}

		var err error
		dir, err = ioutil.TempDir("", "")
		Expect(err).ToNot(HaveOccurred())
		conf := &services.HubConfig{
			HubToAgentPort: port,
			StateDir:       dir,
		}

		clusterPair = testutils.CreateSampleClusterPair()
		testExecutor = &testhelper.TestExecutor{}
		clusterPair.OldCluster.Executor = testExecutor

		// Mock so statusConversion doesn't need to wait 3 seconds before erroring out.
		mockDialer := func(ctx context.Context, target string, opts ...grpc.DialOption) (*grpc.ClientConn, error) {
			return nil, errors.New("grpc dial err")
		}

		cm = testutils.NewMockChecklistManager()
		// XXX this is wrong
		cm.LoadSteps([]upgradestatus.Step{
			{upgradestatus.CONFIG, pb.UpgradeSteps_CHECK_CONFIG, nil},
			{upgradestatus.INIT_CLUSTER, pb.UpgradeSteps_PREPARE_INIT_CLUSTER, nil},
			{upgradestatus.SEGINSTALL, pb.UpgradeSteps_SEGINSTALL, nil},
			{upgradestatus.SHUTDOWN_CLUSTERS, pb.UpgradeSteps_STOPPED_CLUSTER, nil},
			{upgradestatus.CONVERT_MASTER, pb.UpgradeSteps_MASTERUPGRADE, nil},
			{upgradestatus.START_AGENTS, pb.UpgradeSteps_PREPARE_START_AGENTS, nil},
			{upgradestatus.SHARE_OIDS, pb.UpgradeSteps_SHARE_OIDS, nil},
			{upgradestatus.VALIDATE_START_CLUSTER, pb.UpgradeSteps_VALIDATE_START_CLUSTER, nil},
			{upgradestatus.CONVERT_PRIMARY, pb.UpgradeSteps_CONVERT_PRIMARIES, nil},
			{upgradestatus.RECONFIGURE_PORTS, pb.UpgradeSteps_RECONFIGURE_PORTS, nil},
		})

		hub = services.NewHub(clusterPair, mockDialer, conf, cm)
	})

	AfterEach(func() {
		utils.System = utils.InitializeSystemFunctions()
		os.RemoveAll(dir)
	})

	It("responds with the statuses of the steps based on checklist state", func() {
		for _, name := range []string{upgradestatus.CONFIG, upgradestatus.SEGINSTALL, upgradestatus.START_AGENTS} {
			step := cm.StepWriter(name)
			step.MarkInProgress()
			step.MarkComplete()
		}

		step := cm.StepWriter(upgradestatus.SHARE_OIDS)
		step.MarkInProgress()
		step.MarkFailed()

		resp, err := hub.StatusUpgrade(nil, &pb.StatusUpgradeRequest{})
		Expect(err).To(BeNil())

		Expect(resp.ListOfUpgradeStepStatuses).To(ConsistOf(
			[]*pb.UpgradeStepStatus{
				{
					Step:   pb.UpgradeSteps_CHECK_CONFIG,
					Status: pb.StepStatus_COMPLETE,
				}, {
					Step:   pb.UpgradeSteps_PREPARE_INIT_CLUSTER,
					Status: pb.StepStatus_PENDING,
				}, {
					Step:   pb.UpgradeSteps_SEGINSTALL,
					Status: pb.StepStatus_COMPLETE,
				}, {
					Step:   pb.UpgradeSteps_STOPPED_CLUSTER,
					Status: pb.StepStatus_PENDING,
				}, {
					Step:   pb.UpgradeSteps_MASTERUPGRADE,
					Status: pb.StepStatus_PENDING,
				}, {
					Step:   pb.UpgradeSteps_PREPARE_START_AGENTS,
					Status: pb.StepStatus_COMPLETE,
				}, {
					Step:   pb.UpgradeSteps_SHARE_OIDS,
					Status: pb.StepStatus_FAILED,
				}, {
					Step:   pb.UpgradeSteps_VALIDATE_START_CLUSTER,
					Status: pb.StepStatus_PENDING,
				}, {
					Step:   pb.UpgradeSteps_CONVERT_PRIMARIES,
					Status: pb.StepStatus_PENDING,
				}, {
					Step:   pb.UpgradeSteps_RECONFIGURE_PORTS,
					Status: pb.StepStatus_PENDING,
				},
			}))
	})

	// TODO: Get rid of these tests once full rewritten test coverage exists
	Describe("creates a reply", func() {
		It("sends status messages under good condition", func() {
			formulatedResponse, err := hub.StatusUpgrade(nil, fakeStatusUpgradeRequest)
			Expect(err).To(BeNil())
			countOfStatuses := len(formulatedResponse.GetListOfUpgradeStepStatuses())
			Expect(countOfStatuses).ToNot(BeZero())
		})

		It("reports that prepare start-agents is pending", func() {
			utils.System.FilePathGlob = func(string) ([]string, error) {
				return []string{"somefile"}, nil
			}

			var fakeStatusUpgradeRequest *pb.StatusUpgradeRequest

			formulatedResponse, err := hub.StatusUpgrade(nil, fakeStatusUpgradeRequest)
			Expect(err).To(BeNil())

			stepStatuses := formulatedResponse.GetListOfUpgradeStepStatuses()

			var stepStatusSaved *pb.UpgradeStepStatus
			for _, stepStatus := range stepStatuses {

				if stepStatus.GetStep() == pb.UpgradeSteps_PREPARE_START_AGENTS {
					stepStatusSaved = stepStatus
				}
			}
			Expect(stepStatusSaved.GetStep()).ToNot(BeZero())
			Expect(stepStatusSaved.GetStatus()).To(Equal(pb.StepStatus_PENDING))
		})

		It("reports that prepare start-agents is running and then complete", func() {
			// TODO this is no longer a really useful test.
			pollStatusUpgrade := func() *pb.UpgradeStepStatus {
				response, _ := hub.StatusUpgrade(nil, &pb.StatusUpgradeRequest{})

				stepStatuses := response.GetListOfUpgradeStepStatuses()

				var stepStatusSaved *pb.UpgradeStepStatus
				for _, stepStatus := range stepStatuses {

					if stepStatus.GetStep() == pb.UpgradeSteps_PREPARE_START_AGENTS {
						stepStatusSaved = stepStatus
					}
				}
				return stepStatusSaved
			}

			step := cm.StepWriter("start-agents")
			step.MarkInProgress()

			status := pollStatusUpgrade()
			Expect(status.GetStep()).ToNot(BeZero())
			Expect(status.GetStatus()).To(Equal(pb.StepStatus_RUNNING))

			step.MarkComplete()

			status = pollStatusUpgrade()
			Expect(status.GetStep()).ToNot(BeZero())
			Expect(status.GetStatus()).To(Equal(pb.StepStatus_COMPLETE))
		})
	})

	Describe("Status of PrepareNewClusterConfig", func() {
		It("marks this step pending if there's no new cluster config file", func() {
			utils.System.Stat = func(filename string) (os.FileInfo, error) {
				return nil, errors.New("cannot find file") /* This is normally a PathError */
			}
			status := services.GetPrepareNewClusterConfigStatus(dir)
			Expect(status).To(Equal(pb.StepStatus_PENDING))
		})

		It("marks this step complete if there is a new cluster config file", func() {
			utils.System.Stat = func(filename string) (os.FileInfo, error) {
				return nil, nil
			}

			status := services.GetPrepareNewClusterConfigStatus(dir)
			Expect(status).To(Equal(pb.StepStatus_COMPLETE))
		})
	})

	Describe("Status of ShutdownClusters", func() {
		It("We're sending the status of shutdown clusters", func() {
			formulatedResponse, err := hub.StatusUpgrade(nil, fakeStatusUpgradeRequest)
			Expect(err).To(BeNil())
			countOfStatuses := len(formulatedResponse.GetListOfUpgradeStepStatuses())
			Expect(countOfStatuses).ToNot(BeZero())
			found := false
			for _, v := range formulatedResponse.GetListOfUpgradeStepStatuses() {
				if pb.UpgradeSteps_STOPPED_CLUSTER == v.Step {
					found = true
				}
			}
			Expect(found).To(Equal(true))
		})
	})
})

func setStateFile(dir string, step string, state string) {
	err := os.MkdirAll(filepath.Join(dir, step), os.ModePerm)
	Expect(err).ToNot(HaveOccurred())

	f, err := os.Create(filepath.Join(dir, step, state))
	Expect(err).ToNot(HaveOccurred())
	f.Close()
}
